+++ lib/pages/practice/practice_page_manager.dart
@@ -19,7 +19,6 @@ class PracticePageManager {
   final promptNotifier = ValueNotifier<String>('');
   final answerNotifier = ValueNotifier<TextSpan>(const TextSpan());
   final isShowingAnswerNotifier = ValueNotifier<bool>(false);
-  final showHintsNotifier = ValueNotifier<bool>(true);
 
   late void Function() _onFinished;
   late List<Verse> _verses;
@@ -37,8 +36,6 @@ class PracticePageManager {
     _onFinished = onFinished;
     promptNotifier.value = _verses[0].prompt;
     countNotifier.value = _verses.length.toString();
-    final showHints = await userSettings.getShowHints();
-    showHintsNotifier.value = showHints;
   }
 
   void show() {